remove conformer ctc; minor fixes in RNN-T
